REFACTOR: Remove unused scripts
Some old versions of the scripts for the runtime and pharmacogenetic tests are still stored in the repo. Since we are unable to run the competing methods in Code Ocean, there is little point in producing a runtime graph. The pharmacogenetic script has been refactored and placed into /code/tests.

Remove the old and unused runtime and pharmacogenetic scripts.