CHORE: Pin new package versions
Packages that were no longer supported had wildcards inserted in their
versions so they could be more reliably run in the future. However, Code
Ocean requires all packages to have their versions pinned.

Pin the newer versions of the previously out-of-date packages. Note that
it is likely that in the future these or other packages may need to be
reset to the wildcard versions if these pinned versions become
unsupported. In this case they will then need to be repinned to the new
versions.