FIX: Increase attribute samples to 200
Sampling to create the attribute graph is limited to 100 to reduce the
run time of the script. However, this makes the graph a little too
noisy.

Increase the sampling to 200. This may increase attribute's run time to
over an hour, so it can't be increased much more than this without
causing the full test set to take longer than Code Ocean's limit of 10
hours. If greater precision is desired for local reproduction,
increasing the max_samples to 2500 will produce similar precision to
that in the paper.